requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
Werkzeug==2.3.7
python-socketio==5.10.0
eventlet==0.33.3
PyJWT==2.8.0 